+++ packages/jest-worker/src/types.ts
@@ -79,6 +79,7 @@ export type FarmOptions = {
   computeWorkerKey?: (method: string, ...args: Array<unknown>) => string | null;
   exposedMethods?: ReadonlyArray<string>;
   forkOptions?: ForkOptions;
+  workerSchedulingPolicy?: 'round-robin' | 'first-come';
   resourceLimits?: ResourceLimits;
   setupArgs?: Array<unknown>;
   maxRetries?: number;